requirements.txt
torch==2.3.0
transformers==4.41.0
datasets==2.19.1
accelerate==2.30.1
trl==0.8.6
tokenizers==0.19.1
hydra-core==1.3.2
pandas==2.2.2
scikit-learn==1.4.2
tqdm==4.66.1
